pass most of libc-test(static) except pthread

+++ include/common.h
@@ -41,7 +41,8 @@ typedef unsigned int mode_t;
 // used in struct kstat
 typedef unsigned long int dev_t;
 typedef unsigned long int ino_t;
-typedef unsigned long int nlink_t;
+// typedef unsigned long int nlink_t;
+typedef unsigned int nlink_t;
 typedef unsigned int uid_t;
 typedef unsigned int gid_t;
 typedef long int off_t;
